+++ src/cz/vendasky/FractionalTree.java
@@ -7,7 +7,7 @@ public class FractionalTree {
     private Fraction actualFraction;
     private String path;
 
-    private FractionalTree(int startNumerator, int startDenominator) {
+    private FractionalTree(long startNumerator, long startDenominator) {
         this.actualFraction = new Fraction(startNumerator, startDenominator);
         this.path = "";
     }
@@ -53,8 +53,8 @@ public class FractionalTree {
     }
 
     public static FractionalTree createFractionalTree(String line) {
-        int startNumerator = Integer.parseInt(line.substring(0, line.indexOf(" ")));
-        int startDenominator = Integer.parseInt(line.substring(line.indexOf(" ") + 1));
+        long startNumerator = Long.parseLong(line.substring(0, line.indexOf(" ")));
+        long startDenominator = Long.parseLong(line.substring(line.indexOf(" ") + 1));
         return new FractionalTree(startNumerator, startDenominator);
     }
